refactor(Services.Formatter): nest concrete formatter types under singular Formatter namespace; split into per-type files
### Type relocation (16 types nested + sub-namespaced)

  MarkdownSearchResultFormatter -> Services.Formatter.Markdown
  TextSearchResultFormatter     -> Services.Formatter.Text
  JSONSearchResultFormatter     -> Services.Formatter.JSON
  HIGMarkdownFormatter          -> Services.Formatter.HIG.Markdown
  HIGTextFormatter              -> Services.Formatter.HIG.Text
  HIGJSONFormatter              -> Services.Formatter.HIG.JSON
  FrameworksMarkdownFormatter   -> Services.Formatter.Frameworks.Markdown
  FrameworksTextFormatter       -> Services.Formatter.Frameworks.Text
  FrameworksJSONFormatter       -> Services.Formatter.Frameworks.JSON
  UnifiedSearchInput            -> Services.Formatter.Unified.Input
  UnifiedSearchMarkdownFormatter -> Services.Formatter.Unified.Markdown
  UnifiedSearchTextFormatter    -> Services.Formatter.Unified.Text
  UnifiedSearchJSONFormatter    -> Services.Formatter.Unified.JSON
  FooterKind                    -> Services.Formatter.Footer.Kind
  FooterItem                    -> Services.Formatter.Footer.Item
  FooterProvider                -> Services.Formatter.Footer.Provider
  SearchFooter                  -> Services.Formatter.Footer.Search
  FooterFormattable             -> Services.Formatter.Footer.Formattable
  MarkdownFooterFormatter       -> Services.Formatter.Footer.Markdown
  TextFooterFormatter           -> Services.Formatter.Footer.Text
  SearchResultFormatConfig      -> Services.Formatter.Config

All drop the redundant 'Formatter' / 'Search' / etc. suffixes — the
namespace path already conveys context.

### One type per file (task #21)

Was 3 multi-type files (FooterFormatter.swift 7 types,
MarkdownFormatter.swift 5 types, UnifiedSearchFormatters.swift 2 types)
plus 2 partially-mixed files (JSONSearchResultFormatter.swift had 2,
HIGTextFormatter.swift had 2, TextSearchResultFormatter.swift had 2,
ResultFormatter.swift had 2 + String extensions).

Now 22 single-type files in Sources/Services/Formatters/:

  Services.Formatter.Result.swift              (protocol)
  Services.Formatter.Config.swift              (was SearchResultFormatConfig)
  Services.Formatter.StringExtensions.swift    (truncated / cleanedForDisplay / etc.)
  Services.Formatter.TeaserResults.swift       (unchanged)
  Services.Formatter.Markdown.swift
  Services.Formatter.Text.swift
  Services.Formatter.JSON.swift
  Services.Formatter.HIG.{Markdown,Text,JSON}.swift
  Services.Formatter.Frameworks.{Markdown,Text,JSON}.swift
  Services.Formatter.Unified.{Input,Markdown,Text,JSON}.swift
  Services.Formatter.Footer.{Kind,Item,Provider,Search,Formattable,Markdown,Text}.swift

### External caller sweep

All 16+ files across Sources/ and Tests/ swept to the qualified path:

  MarkdownSearchResultFormatter   -> Services.Formatter.Markdown
  HIGTextFormatter                -> Services.Formatter.HIG.Text
  Services.Formatter.HIGJSONFormatter -> Services.Formatter.HIG.JSON
  Services.Formatter.FrameworksTextFormatter -> Services.Formatter.Frameworks.Text
  ...etc

### Verification

- xcrun swift build (clean): 0 warnings, 0 errors
- xcrun swift test (clean):  1300/1300 pass

Closes the Services.Formatter restructure that was sketched in #421.
