+++ Packages/Sources/Services/Formatters/Services.Formatter.Footer.Formattable.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+// MARK: - Footer Formattable Protocol
+
+extension Services.Formatter.Footer {
+    /// Protocol for formatting footer content
+    public protocol Formattable {
+        func format(_ items: [Item]) -> String
+    }
+}
