+++ Packages/Sources/Services/Formatters/Services.Formatter.Footer.Provider.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+// MARK: - Footer Provider Protocol
+
+extension Services.Formatter.Footer {
+    /// Protocol for types that can provide footer content
+    public protocol Provider: Sendable {
+        /// Generate footer items for this context
+        func makeFooter() -> [Item]
+    }
+}
