+++ Packages/Sources/Services/Formatters/Services.Formatter.Result.swift
@@ -1,0 +1,13 @@
+import Foundation
+import Search
+import SharedCore
+
+// MARK: - Result Formatter Protocol
+
+extension Services.Formatter {
+    /// Protocol for formatting search results to different output formats
+    public protocol Result {
+        associatedtype Input
+        func format(_ input: Input) -> String
+    }
+}
