+++ Packages/Sources/Services/Formatters/Services.Formatter.Text.swift
@@ -5,18 +5,18 @@ import SharedCore
 
 // MARK: - Text Search Result Formatter
 
-/// Formats search results as plain text for CLI output
 extension Services.Formatter {
-    public struct TextSearchResultFormatter: Services.Formatter.Result {
+    /// Formats search results as plain text for CLI output
+    public struct Text: Result {
         private let query: String
         private let source: String?
-        private let config: Services.Formatter.SearchResultFormatConfig
+        private let config: Services.Formatter.Config
         private let teasers: TeaserResults?
 
         public init(
             query: String,
             source: String? = nil,
-            config: Services.Formatter.SearchResultFormatConfig = .cliDefault,
+            config: Services.Formatter.Config = .cliDefault,
             teasers: TeaserResults? = nil
         ) {
             self.query = query
@@ -75,39 +75,7 @@ extension Services.Formatter {
 
             // Footer: teasers, tips, and guidance
             let searchedSource = source ?? Shared.Constants.SourcePrefix.appleDocs
-            let footer = SearchFooter.singleSource(searchedSource, teasers: teasers)
-            output += footer.formatText()
-
-            return output
-        }
-    }
-}
-
-// MARK: - Frameworks Text Formatter
-
-/// Formats framework list as plain text for CLI output
-extension Services.Formatter {
-    public struct FrameworksTextFormatter: Services.Formatter.Result {
-        private let totalDocs: Int
-
-        public init(totalDocs: Int) {
-            self.totalDocs = totalDocs
-        }
-
-        public func format(_ frameworks: [String: Int]) -> String {
-            if frameworks.isEmpty {
-                return "No frameworks found. Run 'cupertino save' to build the index."
-            }
-
-            var output = "Available Frameworks (\(frameworks.count) total, \(totalDocs) documents):\n\n"
-
-            // Sort by document count (descending)
-            for (framework, count) in frameworks.sorted(by: { $0.value > $1.value }) {
-                output += "  \(framework): \(count) documents\n"
-            }
-
-            // Footer: tips and guidance
-            let footer = SearchFooter.singleSource(Shared.Constants.SourcePrefix.appleDocs)
+            let footer = Services.Formatter.Footer.Search.singleSource(searchedSource, teasers: teasers)
             output += footer.formatText()
 
             return output
